requirements.txt
flask
flask_sock
pyperclip
websockets
mysql-connector-python